backend/service/requirements.txt
fastapi
uvicorn
pydantic
fastapi
uvicorn
pydantic
langchain-community
openai
pandas
pinecone
langchain_openai
python-dotenv
sys